refactor: Journals model 파일 journal로 변경
