+++ frontend/src/models/journal.ts
